requirements.txt
# Child Nutrition Data ETL Pipeline Dependencies
# Core data processing
pandas==2.0.0
numpy==1.24.0

# Snowflake integration
snowflake-connector-python==3.0.0

# Environment and configuration
python-dotenv==1.0.0

# Testing framework
pytest==8.4.2
pytest-cov==4.1.0

# Logging and monitoring
loguru==0.7.0

# Data validation
great-expectations==0.17.0

# Development tools
black==23.7.0
flake8==6.0.0
mypy==1.5.0

# Documentation
sphinx==7.1.0
sphinx-rtd-theme==1.3.0
